requirements.txt
Flask==2.3.0
pyTelegramBotAPI==4.12.0
openai==1.0.0
easyocr==1.6.2
pillow==10.0.0
torch
